+++ include/gas_simulation_app.h
@@ -18,9 +18,9 @@ class IdealGasApp : public ci::app::App {
   void update() override;
   void keyDown(cinder::app::KeyEvent event) override;
 
-  const int kWindowLength = 800;
-  const int kWindowWidth = static_cast<int>(1.6 * kWindowLength);
-  const int kMargin = 80;
+  const size_t kWindowLength = 800;
+  const size_t kWindowWidth = static_cast<int>(1.6 * kWindowLength);
+  const size_t kMargin = 80;
   const ci::Color kBorderColor = ci::Color("white");
 
  private:
